Expose server runner API for embedding
